+++ ec/utils.py
@@ -8,25 +8,27 @@ from modules.exposed import get, static
 from modules.classes import CustomType
 
 __all__ = ['get', 'static', 'custom']
-  
-def custom(validator, converter=None, **Config):
+
+class custom(CustomType):
   """Helps with creating dynamic CustomTypes on the fly.
   
   Args:
-    validator (callable): that acts as the converter / validator.
-    converter (callable):  A callable that converts the given value. Defaults to None.
+    validator (callable): Validates the input.
+    converter (callable): Converts the input. Defaults to None.
+    **Config  (kwargs): The configuration of the CustomType
   """
-  class dynamic(CustomType):
-    def __init__(self):
-      CustomType.__init__(self, **Config)
-      
-    def __call__(self, val):
-      if converter:
-        val = converter(val)
-        
-      if not validator(val):
-        raise ValueError()
-        
-      return val
+  def __init__(self, validator, converter=None, **Config):
+    CustomType.__init__(self, **Config)
     
-  return dynamic()
+    self.validator = validator
+    self.converter = converter
+    
+  def __call__(self, val):
+    if self.converter:
+      val = self.converter(val)
+      
+    if not self.validator(val):
+      raise ValueError()
+      
+    return val
+